chore(runtime): keep liveness broker under line limit
